Don't init logrus in a seperate file
The logging funcationality is broken when crossing compile the
executable file due to some unknown machnisms related to the main
package. So, move `log_hook.go` to `main.go`.
